[ICU-2966] Fix for HP-UX X-Commit-URL: https://ssl.icu-project.org/trac/changeset/12936 Former-commit-id: cf8bd158a8f18561747945b70cf762d10108e703

+++ icu4c/source/i18n/olsontz.cpp
@@ -46,7 +46,7 @@ static double floorDivide(double numerator, double denominator,
 static int32_t floorDivide(double numerator, int32_t denominator,
                            int32_t& remainder) {
     double quotient, rem;
-    quotient = floorDivide(numerator, denominator, rem);
+    quotient = floorDivide(numerator, (double)denominator, rem);
     remainder = (int32_t) rem;
     return (int32_t) quotient;
 }
